+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 * demodata images now have the correct extension
 
+* Fixed issue with channelless data in `Heatmap.draw_on`
 
 ### Changed